// src/app/models/self/cls_MTTopic.ts
export class cls_MTTopicModel {
    constructor() {

    }
    company_code: string = "";
    topic_id: number = 0;
    topic_code: string = "";
    topic_name_th: string = "";
    topic_name_en: string = "";
    topic_type: string = "";
    modified_by: string = "";
    modified_date: string = "";
    flag: boolean = false;
}